pyproject.toml
[project]
name = "potto"
version = "0.1.0"
description = "A starlette application that wraps pygeoapi"
readme = "README.md"
authors = [
    { name = "Ricardo Garcia Silva", email = "ricardo.garcia.silva@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.0",
    "alembic-postgresql-enum>=1.8.0",
    "babel>=2.17.0",
    "cyclopts>=4.2.4",
    "fastapi[standard]>=0.128.0",
    "geoalchemy2>=0.18.1",
    "jinja2>=3.1.6",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pygeoapi>=0.22.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "starlette>=0.50.0",
    "starlette-admin>=0.16.0",
    "starlette-babel>=1.0.3",
    "uvicorn>=0.38.0",
]

[project.scripts]
potto = "potto.__main__:potto_app.meta"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
